refactor: extract shared Placement into homeboy-cli-contract crate
core routing referenced crate::cli_surface::Placement in ~45 places, but
cli_surface is the full clap CLI definition that depends on all of commands.
That made core transitively depend on the CLI layer and blocks extracting
commands into its own crate.

Move Placement (the execution-location enum) plus its inherent routing methods
(allows_local_fallback, requests_lab) into a new leaf crate
homeboy-cli-contract, depending only on clap + serde. core now depends on this
small shared contract instead of cli_surface; cli_surface re-exports Placement
so its own call sites are unchanged. The inherent-impl methods had to move with
the enum (inherent impls must be defined in the type's home crate).

This removes the Placement edge from core -> cli_surface. The remaining
core -> cli_surface references are the full Cli/Commands parser calls in runner
(workload/connection/provider_config), which are a separate, deeper inversion
(core should not parse full CLI argv) tracked as follow-up.

Not published to crates.io; still ships as the single homeboy binary.
cargo check --lib and --lib --tests pass.
