+++ Cargo.toml
@@ -37,6 +37,7 @@ path = "src/bin/bench-audit-self.rs"
 
 [dependencies]
 # Internal workspace crates (path-only, not published)
+homeboy-cli-contract = { path = "crates/homeboy-cli-contract" }
 homeboy-error = { path = "crates/homeboy-error" }
 homeboy-finding = { path = "crates/homeboy-finding" }
 homeboy-output = { path = "crates/homeboy-output" }
