+++ src/core/lab_routing.rs
@@ -29,7 +29,7 @@ pub struct LabRoutingRequest<'a> {
     pub command: Option<runners::LabOffloadCommand>,
     pub normalized_args: &'a [String],
     pub explicit_runner: Option<&'a str>,
-    pub placement: crate::cli_surface::Placement,
+    pub placement: homeboy_cli_contract::Placement,
     pub allow_local_fallback: bool,
     pub allow_dirty_lab_workspace: bool,
     pub skip_deps_hydration: bool,
@@ -728,7 +728,7 @@ mod tests {
             command: None,
             normalized_args: &["homeboy".to_string(), "status".to_string()],
             explicit_runner: None,
-            placement: crate::cli_surface::Placement::Auto,
+            placement: homeboy_cli_contract::Placement::Auto,
             allow_local_fallback: false,
             allow_dirty_lab_workspace: false,
             skip_deps_hydration: false,
@@ -1031,7 +1031,7 @@ mod tests {
                 command: None,
                 normalized_args: &["homeboy".to_string(), "status".to_string()],
                 explicit_runner: None,
-                placement: crate::cli_surface::Placement::Auto,
+                placement: homeboy_cli_contract::Placement::Auto,
                 allow_local_fallback: false,
                 allow_dirty_lab_workspace: false,
                 skip_deps_hydration: false,
