+++ src/core/runner/lab/offload/tests/exec_errors.rs
@@ -282,7 +282,7 @@ fn default_runner_missing_capabilities_fails_without_local_fallback_opt_in() {
             .to_string(),
         vec!["Install Playwright and browser binaries on the runner.".to_string()],
         false,
-        crate::cli_surface::Placement::Auto,
+        homeboy_cli_contract::Placement::Auto,
         &overhead,
     );
 
@@ -325,7 +325,7 @@ fn default_runner_missing_capabilities_can_fallback_with_explicit_opt_in() {
             .to_string(),
         Vec::new(),
         true,
-        crate::cli_surface::Placement::Auto,
+        homeboy_cli_contract::Placement::Auto,
         &overhead,
     )
     .expect("explicit fallback opt-in should allow local run");
@@ -364,7 +364,7 @@ fn plan_records_skipped_auto_offload() {
         command: Some(portable_lab_command("test")),
         normalized_args: &["homeboy".to_string(), "test".to_string()],
         explicit_runner: None,
-        placement: crate::cli_surface::Placement::Local,
+        placement: homeboy_cli_contract::Placement::Local,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
@@ -396,7 +396,7 @@ fn lab_placement_refuses_local_execution_without_lab_contract() {
         command: None,
         normalized_args: &["homeboy".to_string(), "status".to_string()],
         explicit_runner: None,
-        placement: crate::cli_surface::Placement::Lab,
+        placement: homeboy_cli_contract::Placement::Lab,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
@@ -431,7 +431,7 @@ fn lab_placement_refuses_local_only_rig_install_with_actionable_boundary() {
             "./rig-package".to_string(),
         ],
         explicit_runner: None,
-        placement: crate::cli_surface::Placement::Lab,
+        placement: homeboy_cli_contract::Placement::Lab,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
@@ -468,7 +468,7 @@ fn build_runner_error_gives_managed_runner_replacement() {
             "homeboy".to_string(),
         ],
         explicit_runner: Some("homeboy-lab"),
-        placement: crate::cli_surface::Placement::Auto,
+        placement: homeboy_cli_contract::Placement::Auto,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
@@ -514,7 +514,7 @@ fn build_lab_placement_error_gives_managed_runner_replacement() {
             "homeboy".to_string(),
         ],
         explicit_runner: None,
-        placement: crate::cli_surface::Placement::Lab,
+        placement: homeboy_cli_contract::Placement::Lab,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
@@ -562,7 +562,7 @@ fn unsupported_runner_error_guides_tunnel_service_inspection() {
             "wpcom-ai-manual-held".to_string(),
         ],
         explicit_runner: Some("homeboy-lab"),
-        placement: crate::cli_surface::Placement::Auto,
+        placement: homeboy_cli_contract::Placement::Auto,
         allow_local_fallback: false,
         allow_dirty_lab_workspace: false,
         skip_deps_hydration: false,
