+++ src/core/runner/lab/offload/workspace_stage.rs
@@ -2137,7 +2137,7 @@ mod tests {
                 command: None,
                 normalized_args: &args,
                 explicit_runner: Some("lab-controller-bundle-stage"),
-                placement: crate::cli_surface::Placement::Auto,
+                placement: homeboy_cli_contract::Placement::Auto,
                 allow_local_fallback: false,
                 allow_dirty_lab_workspace: false,
                 skip_deps_hydration: false,
@@ -2289,7 +2289,7 @@ mod tests {
                 command: None,
                 normalized_args: &args,
                 explicit_runner: Some("lab-managed-worktree-stage"),
-                placement: crate::cli_surface::Placement::Lab,
+                placement: homeboy_cli_contract::Placement::Lab,
                 allow_local_fallback: false,
                 allow_dirty_lab_workspace: false,
                 skip_deps_hydration: false,
